Model and Dto added
