+++ main-service/src/main/java/ru/practicum/ewm/category/dto/CategoryDto.java
@@ -1,6 +1,6 @@
-package ru.practicum.ewm.compilation;
+package ru.practicum.ewm.category.dto;
 
-import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotBlank;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -11,14 +11,12 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @AllArgsConstructor
 @NoArgsConstructor
-public class CompilationDto {
+public class CategoryDto {
 
-    @NotNull
     Long id;
 
-    Boolean pinned;
+    @NotBlank
 
-    @NotNull
-    String title;
+    String name;
 
 }
