+++ main-service/src/main/java/ru/practicum/ewm/category/dto/NewCategoryDto.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewm.compilation;
+package ru.practicum.ewm.category.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.AccessLevel;
@@ -7,19 +7,11 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.FieldDefaults;
 
-import java.util.List;
-
 @Data
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @AllArgsConstructor
 @NoArgsConstructor
-public class NewCompilationDto {
-
-    List<Long> events;
-
-    Boolean pinned;
-
+public class NewCategoryDto {
     @NotBlank
-    String title;
-
+    String name;
 }
