+++ main-service/src/main/java/ru/practicum/ewm/user/dto/UserDto.java
@@ -1,26 +1,26 @@
-package ru.practicum.ewm.user;
+package ru.practicum.ewm.user.dto;
 
-import jakarta.persistence.*;
 import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
 import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 import lombok.experimental.FieldDefaults;
 
 @Data
-@Entity
-@Table(name = "users")
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class User {
+@AllArgsConstructor
+@NoArgsConstructor
+public class UserDto {
 
+    @NotBlank
     @Email
-    @Column(name = "email")
     String email;
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
     Long id;
 
-    @Column(name = "name")
+    @NotBlank
     String name;
 
 }
